requirements.txt
streamlit>=1.22.0
pandas>=1.5.3
numpy>=1.24.3
scikit-learn>=1.2.2
nltk>=3.8.1
Pillow>=9.5.0
requests>=2.29.0
matplotlib>=3.7.1
seaborn>=0.12.2